:art: removed unnecessary components - also, refactored a couple things in landing, game cover, app and game components
