+++ frontend/src/pages/app/App.js
@@ -1,12 +1,17 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
+import { Container } from "semantic-ui-react";
 import Landing from "../landing/Landing";
 
 class App extends Component {
   render() {
     if (this.props.auth.isAuthenticated) {
-      return <h1>Welcome</h1>;
+      return (
+        <Container>
+          <h1>Welcome</h1>
+        </Container>
+      );
     } else if (this.props.auth.isLoading) {
       return null;
     } else {
