--- frontend/src/pages/landing/components/headline/Headline.css
@@ -1,15 +1,0 @@
-.landing-header {
-  text-align: center;
-}
-
-.landing-header a {
-  color: #def;
-}
-
-.landing-header a:hover {
-  color: #fff !important;
-}
-
-.landing-header .button {
-  text-shadow: 0 0 7px rgba(0, 0, 0, 0.45) !important;
-}
